requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
